# lib/api-zod/src/generated/types/faucetRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Stellar DLMM Liquidity Protocol API
 * OpenAPI spec version: 0.1.0
 */

export interface FaucetRequest {
  /** Destination Stellar account (G...) that already has a USDC trustline */
  address: string;
}