# lib/api-zod/src/generated/types/positionEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Stellar DLMM Liquidity Protocol API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionEventAction } from './positionEventAction';

export interface PositionEvent {
  txHash: string;
  timestamp: string;
  address: string;
  action: PositionEventAction;
  binId: number;
  /** Raw stroops (i128), as a string to avoid precision loss */
  amountX: string;
  /** Raw stroops (i128), as a string to avoid precision loss */
  amountY: string;
  /** Raw LP shares (i128), as a string to avoid precision loss */
  shares: string;
}